combine error files in gophercloud pkg
